+++ onlyoffice-nuxeo-core/src/main/java/org/onlyoffice/constants/SettingsConstants.java
@@ -23,5 +23,6 @@ public class SettingsConstants {
     public static final String SETTINGS_FILE = "onlyoffice.properties";
     public static final String DOC_SERVER_URL = "onlyoffice.docserv.url";
     public static final String JWT_SECRET = "onlyoffice.jwt.secret";
+    public static final String JWT_HEADER = "onlyoffice.jwt.header";
 
 }